backend/pyproject.toml
[project]
name = "integration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "cryptography>=46.0.5",
    "fastapi[standard]>=0.135.1",
    "greenlet>=3.3.2",
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "sqlmodel>=0.0.37",
]
